feat: Update package name and logo for Jebao Dosing Pump MD 4.4
This commit updates the package name and logo in the README file to reflect the changes made for the Jebao Dosing Pump MD 4.4 device. The package name is now "jebao-dosing-pump-md-4.4" and the logo image has been replaced with "dosing-pump-jebao-md-44.png". Additionally, the badges in the README have been updated accordingly.

A special thanks to Apollon77, whose work laid the foundation for this current project. [His work](https://github.com/Apollon77/node-ph803w)

Changelog:
- Updated package name and version
- Replaced logo image with new one
- Updated badges in README